requirements.txt
flask
langgraph
langchain-google-genai
langchain
langchain-community
PyPDF2
python-dotenv
google-generativeai